timer + other stuff update

+++ projects/libSysCall/src/init.c
@@ -303,7 +303,7 @@ long unsigned int getTime()
     
     assert(seL4_GetMR(0) == SysCall_GetTime);
     
-    return (int) seL4_GetMR(1);
+    return  seL4_GetMR(1);
 }
 
 static int _Sleep( unsigned long l , SleepUnit unit)
